refactor: create cloud-share feature directory (Phase 5g)
Move cloud sharing components, hooks, and utils to features/cloud-share/
with barrel exports. Create backward-compatible re-export stubs at old
locations for gradual migration.

Components: ShareModal, CloudShareTab, ShareButton, SharedLayoutBanner,
SharedLayoutImporter
Hooks: useCloudShare, useCloudShareAutoSync, useOwnedShareSync,
useSharedWithMe
Utils: cloudShare (formatShareDate)

Update test mock paths for relocated modules.
