+++ src/features/cloud-share/components/index.ts
@@ -1,0 +1,6 @@
+// Barrel export for cloud-share components
+export { CloudShareTab } from './CloudShareTab';
+export { ShareButton } from './ShareButton';
+export { SharedLayoutBanner } from './SharedLayoutBanner';
+export { SharedLayoutImporter } from './SharedLayoutImporter';
+export { ShareModal } from './ShareModal';
