+++ src/features/cloud-share/hooks/index.ts
@@ -1,0 +1,5 @@
+// Barrel export for cloud-share hooks
+export { useCloudShare, type CloudShareStatus } from './useCloudShare';
+export { useCloudShareAutoSync } from './useCloudShareAutoSync';
+export { useOwnedShareSync } from './useOwnedShareSync';
+export { useSharedWithMe, type SharedWithMeStatus } from './useSharedWithMe';
