+++ src/features/cloud-share/utils/index.ts
@@ -1,0 +1,2 @@
+// Barrel export for cloud-share utils
+export { formatShareDate } from './cloudShare';
